Revert "New --archive flag saves a configuration to disk in ready-to-upload format (#232)"
This reverts commit 5b8f8937fefe1389dfd1f9e6bb7aa6c9c6e1ca2d.

+++ src/fn/delete-all-forms.js
@@ -1,8 +1,12 @@
-const formsList = require('../lib/forms-list');
+const skipFn = require('../lib/skip-fn');
 const pouch = require('../lib/db');
+const formsList = require('../lib/forms-list');
 
-module.exports = () => {
-  const db = pouch();
+module.exports = (projectDir, couchUrl) => {
+  if(!couchUrl) return skipFn('no couch URL set');
+
+  const db = pouch(couchUrl);
+
   return formsList(db, { include_docs: true })
     .then(res => res.rows)
     .then(forms => Promise.all(forms.map(f => db.remove(f.doc))));
